refactor(git): sync upstream PR and worktree handling
- improve GitHub PR detection and fork upstream setup
- update PR thread worktree flow for cross-repo branches
- align compatibility and upstream merge guidance
