+++ apps/server/src/git/Layers/GitManager.ts
@@ -1,1534 +1,18 @@
-import { randomUUID } from "node:crypto";
-import { realpathSync } from "node:fs";
+import { Effect, Layer } from "effect";
 
-import { Effect, FileSystem, Layer, Path } from "effect";
-import {
-  type GitActionProgressEvent,
-  GitActionProgressPhase,
-  type GitHubIssueLink,
-  type ModelSelection,
-} from "@t3tools/contracts";
-import {
-  buildManagedWorktreeBranchName,
-  resolveAutoFeatureBranchName,
-  sanitizeBranchFragment,
-  sanitizeFeatureBranchName,
-} from "@t3tools/shared/git";
-
-import { GitManagerError } from "../Errors.ts";
-import {
-  buildBlockedPullDetail,
-  buildBlockedPushDetail,
-  isProtectedRemoteName,
-  resolvePublishRemoteName,
-  selectPrimaryRemoteName,
-} from "../gitPolicy.ts";
-import {
-  DEFAULT_GITHUB_HOSTNAME,
-  parseGitHubRepoSelector,
-  resolveGitHubRepositorySelector,
-} from "../gitHubRepository.ts";
-import {
-  GitManager,
-  type GitActionProgressReporter,
-  type GitManagerShape,
-  type GitRunStackedActionOptions,
-} from "../Services/GitManager.ts";
-import { GitCore } from "../Services/GitCore.ts";
-import { GitHubCli } from "../Services/GitHubCli.ts";
-import { TextGeneration } from "../Services/TextGeneration.ts";
-import { ServerSettingsService } from "../../serverSettings.ts";
-
-type StripProgressContext<T> = T extends any ? Omit<T, "actionId" | "cwd" | "action"> : never;
-type GitActionProgressPayload = StripProgressContext<GitActionProgressEvent>;
-
-interface OpenPrInfo {
-  number: number;
-  title: string;
-  url: string;
-  baseRefName: string;
-  headRefName: string;
-}
-
-interface PullRequestInfo extends OpenPrInfo {
-  state: "open" | "closed" | "merged";
-  updatedAt: string | null;
-}
-
-interface ResolvedPullRequest {
-  number: number;
-  title: string;
-  url: string;
-  baseBranch: string;
-  headBranch: string;
-  state: "open" | "closed" | "merged";
-}
-
-interface PullRequestHeadRemoteInfo {
-  isCrossRepository?: boolean;
-  headRepositoryNameWithOwner?: string | null;
-  headRepositoryOwnerLogin?: string | null;
-}
-
-interface BranchHeadContext {
-  localBranch: string;
-  headBranch: string;
-  headSelectors: ReadonlyArray<string>;
-  preferredHeadSelector: string;
-  remoteName: string | null;
-  headRepositoryNameWithOwner: string | null;
-  headRepositoryOwnerLogin: string | null;
-  isCrossRepository: boolean;
-}
-
-function parseRepositoryNameFromPullRequestUrl(url: string): string | null {
-  const trimmed = url.trim();
-  const match = /^https:\/\/github\.com\/[^/]+\/([^/]+)\/pull\/\d+(?:\/.*)?$/i.exec(trimmed);
-  const repositoryName = match?.[1]?.trim() ?? "";
-  return repositoryName.length > 0 ? repositoryName : null;
-}
-
-function resolveHeadRepositoryNameWithOwner(
-  pullRequest: ResolvedPullRequest & PullRequestHeadRemoteInfo,
-): string | null {
-  const explicitRepository = pullRequest.headRepositoryNameWithOwner?.trim() ?? "";
-  if (explicitRepository.length > 0) {
-    return explicitRepository;
-  }
-
-  if (!pullRequest.isCrossRepository) {
-    return null;
-  }
-
-  const ownerLogin = pullRequest.headRepositoryOwnerLogin?.trim() ?? "";
-  const repositoryName = parseRepositoryNameFromPullRequestUrl(pullRequest.url);
-  if (ownerLogin.length === 0 || !repositoryName) {
-    return null;
-  }
-
-  return `${ownerLogin}/${repositoryName}`;
-}
-
-function resolvePullRequestWorktreeLocalBranchName(
-  pullRequest: ResolvedPullRequest & PullRequestHeadRemoteInfo,
-): string {
-  if (!pullRequest.isCrossRepository) {
-    return pullRequest.headBranch;
-  }
-
-  const sanitizedHeadBranch = sanitizeBranchFragment(pullRequest.headBranch).trim();
-  const suffix = sanitizedHeadBranch.length > 0 ? sanitizedHeadBranch : "head";
-  return buildManagedWorktreeBranchName(suffix, `pr-${pullRequest.number}`);
-}
-
-function parseRepositoryOwnerLogin(nameWithOwner: string | null): string | null {
-  const trimmed = nameWithOwner?.trim() ?? "";
-  if (trimmed.length === 0) {
-    return null;
-  }
-  const [ownerLogin] = trimmed.split("/");
-  const normalizedOwnerLogin = ownerLogin?.trim() ?? "";
-  return normalizedOwnerLogin.length > 0 ? normalizedOwnerLogin : null;
-}
-
-function withOptionalRepo<T extends object>(input: T, repo: string | null): T & { repo?: string } {
-  if (repo === null) {
-    return input;
-  }
-
-  return { ...input, repo };
-}
-
-function escapeRegExp(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function buildIssueClosingReference(issueLink: GitHubIssueLink): string {
-  return `Closes ${issueLink.repoNameWithOwner}#${issueLink.number}`;
-}
-
-function bodyAlreadyClosesIssue(body: string, issueLink: GitHubIssueLink): boolean {
-  const issueNumber = String(issueLink.number);
-  const qualifiedReference = `${issueLink.repoNameWithOwner}#${issueNumber}`;
-  const closingVerbPattern = String.raw`(?:close|closes|closed|fix|fixes|fixed|resolve|resolves|resolved)`;
-  const issueReferencePattern = String.raw`(?:${escapeRegExp(qualifiedReference)}|#${escapeRegExp(issueNumber)})`;
-  const pattern = new RegExp(`${closingVerbPattern}\\s+${issueReferencePattern}`, "i");
-  return pattern.test(body);
-}
-
-function appendIssueClosingFooter(body: string, issueLink: GitHubIssueLink | null): string {
-  const trimmedBody = body.trimEnd();
-  if (!issueLink || bodyAlreadyClosesIssue(trimmedBody, issueLink)) {
-    return trimmedBody;
-  }
-
-  const closingReference = buildIssueClosingReference(issueLink);
-  return trimmedBody.length > 0 ? `${trimmedBody}\n\n${closingReference}` : closingReference;
-}
-
-function parsePullRequestList(raw: unknown): PullRequestInfo[] {
-  if (!Array.isArray(raw)) return [];
-
-  const parsed: PullRequestInfo[] = [];
-  for (const entry of raw) {
-    if (!entry || typeof entry !== "object") continue;
-    const record = entry as Record<string, unknown>;
-    const number = record.number;
-    const title = record.title;
-    const url = record.url;
-    const baseRefName = record.baseRefName;
-    const headRefName = record.headRefName;
-    const state = record.state;
-    const mergedAt = record.mergedAt;
-    const updatedAt = record.updatedAt;
-    if (typeof number !== "number" || !Number.isInteger(number) || number <= 0) {
-      continue;
-    }
-    if (
-      typeof title !== "string" ||
-      typeof url !== "string" ||
-      typeof baseRefName !== "string" ||
-      typeof headRefName !== "string"
-    ) {
-      continue;
-    }
-
-    let normalizedState: "open" | "closed" | "merged";
-    if ((typeof mergedAt === "string" && mergedAt.trim().length > 0) || state === "MERGED") {
-      normalizedState = "merged";
-    } else if (state === "OPEN" || state === undefined || state === null) {
-      normalizedState = "open";
-    } else if (state === "CLOSED") {
-      normalizedState = "closed";
-    } else {
-      continue;
-    }
-
-    parsed.push({
-      number,
-      title,
-      url,
-      baseRefName,
-      headRefName,
-      state: normalizedState,
-      updatedAt: typeof updatedAt === "string" && updatedAt.trim().length > 0 ? updatedAt : null,
-    });
-  }
-  return parsed;
-}
-
-function gitManagerError(operation: string, detail: string, cause?: unknown): GitManagerError {
-  return new GitManagerError({
-    operation,
-    detail,
-    ...(cause !== undefined ? { cause } : {}),
-  });
-}
-
-function limitContext(value: string, maxChars: number): string {
-  if (value.length <= maxChars) return value;
-  return `${value.slice(0, maxChars)}\n\n[truncated]`;
-}
-
-function sanitizeCommitMessage(generated: {
-  subject: string;
-  body: string;
-  branch?: string | undefined;
-}): {
-  subject: string;
-  body: string;
-  branch?: string | undefined;
-} {
-  const rawSubject = generated.subject.trim().split(/\r?\n/g)[0]?.trim() ?? "";
-  const subject = rawSubject.replace(/[.]+$/g, "").trim();
-  const safeSubject = subject.length > 0 ? subject.slice(0, 72).trimEnd() : "Update project files";
-  return {
-    subject: safeSubject,
-    body: generated.body.trim(),
-    ...(generated.branch !== undefined ? { branch: generated.branch } : {}),
-  };
-}
-
-interface CommitAndBranchSuggestion {
-  subject: string;
-  body: string;
-  branch?: string | undefined;
-  commitMessage: string;
-}
-
-function formatCommitMessage(subject: string, body: string): string {
-  const trimmedBody = body.trim();
-  if (trimmedBody.length === 0) {
-    return subject;
-  }
-  return `${subject}\n\n${trimmedBody}`;
-}
-
-function parseCustomCommitMessage(raw: string): { subject: string; body: string } | null {
-  const normalized = raw.replace(/\r\n/g, "\n").trim();
-  if (normalized.length === 0) {
-    return null;
-  }
-
-  const [firstLine, ...rest] = normalized.split("\n");
-  const subject = firstLine?.trim() ?? "";
-  if (subject.length === 0) {
-    return null;
-  }
-
-  return {
-    subject,
-    body: rest.join("\n").trim(),
-  };
-}
-
-function extractBranchFromRef(ref: string): string {
-  const normalized = ref.trim();
-
-  if (normalized.startsWith("refs/remotes/")) {
-    const withoutPrefix = normalized.slice("refs/remotes/".length);
-    const firstSlash = withoutPrefix.indexOf("/");
-    if (firstSlash === -1) {
-      return withoutPrefix.trim();
-    }
-    return withoutPrefix.slice(firstSlash + 1).trim();
-  }
-
-  const firstSlash = normalized.indexOf("/");
-  if (firstSlash === -1) {
-    return normalized;
-  }
-  return normalized.slice(firstSlash + 1).trim();
-}
-
-function appendUnique(values: string[], next: string | null | undefined): void {
-  const trimmed = next?.trim() ?? "";
-  if (trimmed.length === 0 || values.includes(trimmed)) {
-    return;
-  }
-  values.push(trimmed);
-}
-
-function parseRemoteNames(stdout: string): ReadonlyArray<string> {
-  return stdout
-    .split("\n")
-    .map((line) => line.trim())
-    .filter((line) => line.length > 0);
-}
-
-function toStatusPr(pr: PullRequestInfo): {
-  number: number;
-  title: string;
-  url: string;
-  baseBranch: string;
-  headBranch: string;
-  state: "open" | "closed" | "merged";
-} {
-  return {
-    number: pr.number,
-    title: pr.title,
-    url: pr.url,
-    baseBranch: pr.baseRefName,
-    headBranch: pr.headRefName,
-    state: pr.state,
-  };
-}
-
-function normalizePullRequestReference(reference: string): string {
-  const trimmed = reference.trim();
-  const hashNumber = /^#(\d+)$/.exec(trimmed);
-  return hashNumber?.[1] ?? trimmed;
-}
-
-function canonicalizeExistingPath(value: string): string {
-  try {
-    return realpathSync.native(value);
-  } catch {
-    return value;
-  }
-}
-
-function toResolvedPullRequest(pr: {
-  number: number;
-  title: string;
-  url: string;
-  baseRefName: string;
-  headRefName: string;
-  state?: "open" | "closed" | "merged";
-}): ResolvedPullRequest {
-  return {
-    number: pr.number,
-    title: pr.title,
-    url: pr.url,
-    baseBranch: pr.baseRefName,
-    headBranch: pr.headRefName,
-    state: pr.state ?? "open",
-  };
-}
-
-function shouldPreferSshRemote(url: string | null): boolean {
-  if (!url) return false;
-  const trimmed = url.trim();
-  return trimmed.startsWith("git@") || trimmed.startsWith("ssh://");
-}
-
-function toPullRequestHeadRemoteInfo(pr: {
-  isCrossRepository?: boolean;
-  headRepositoryNameWithOwner?: string | null;
-  headRepositoryOwnerLogin?: string | null;
-}): PullRequestHeadRemoteInfo {
-  return {
-    ...(pr.isCrossRepository !== undefined ? { isCrossRepository: pr.isCrossRepository } : {}),
-    ...(pr.headRepositoryNameWithOwner !== undefined
-      ? { headRepositoryNameWithOwner: pr.headRepositoryNameWithOwner }
-      : {}),
-    ...(pr.headRepositoryOwnerLogin !== undefined
-      ? { headRepositoryOwnerLogin: pr.headRepositoryOwnerLogin }
-      : {}),
-  };
-}
+import { ForkGitDomain } from "../Services/ForkGitDomain.ts";
+import { GitManager, type GitManagerShape } from "../Services/GitManager.ts";
 
 export const makeGitManager = Effect.gen(function* () {
-  const gitCore = yield* GitCore;
-  const gitHubCli = yield* GitHubCli;
-  const textGeneration = yield* TextGeneration;
-  const serverSettingsService = yield* ServerSettingsService;
-
-  const resolveRepositorySelector = (cwd: string) =>
-    resolveGitHubRepositorySelector(cwd, gitCore, DEFAULT_GITHUB_HOSTNAME);
-
-  const createProgressEmitter = (
-    input: Parameters<GitManagerShape["runStackedAction"]>[0],
-    options?: GitRunStackedActionOptions,
-  ) => {
-    const actionId = options?.actionId ?? input.actionId;
-
-    const emit = (payload: GitActionProgressPayload) => {
-      if (!options?.progressReporter) {
-        return Effect.void;
-      }
-      return options.progressReporter.publish({
-        ...payload,
-        actionId,
-        cwd: input.cwd,
-        action: input.action,
-      } as GitActionProgressEvent);
-    };
-
-    return {
-      actionId,
-      emit,
-    };
-  };
-
-  const resolveRepositoryContext = (repositoryNameWithOwner: string | null) => ({
-    repositoryNameWithOwner,
-    ownerLogin: parseRepositoryOwnerLogin(repositoryNameWithOwner),
-  });
-
-  const configurePullRequestHeadUpstream = (
-    cwd: string,
-    pullRequest: ResolvedPullRequest & PullRequestHeadRemoteInfo,
-    localBranch = pullRequest.headBranch,
-  ) =>
-    Effect.gen(function* () {
-      const repositoryNameWithOwner = resolveHeadRepositoryNameWithOwner(pullRequest) ?? "";
-      if (repositoryNameWithOwner.length === 0) {
-        return;
-      }
-
-      const cloneUrls = yield* gitHubCli.getRepositoryCloneUrls({
-        cwd,
-        repository: repositoryNameWithOwner,
-      });
-      const originRemoteUrl = yield* gitCore.readConfigValue(cwd, "remote.origin.url");
-      const remoteUrl = shouldPreferSshRemote(originRemoteUrl) ? cloneUrls.sshUrl : cloneUrls.url;
-      const preferredRemoteName =
-        pullRequest.headRepositoryOwnerLogin?.trim() ||
-        repositoryNameWithOwner.split("/")[0]?.trim() ||
-        "fork";
-      const remoteName = yield* gitCore.ensureRemote({
-        cwd,
-        preferredName: preferredRemoteName,
-        url: remoteUrl,
-      });
-
-      yield* gitCore.setBranchUpstream({
-        cwd,
-        branch: localBranch,
-        remoteName,
-        remoteBranch: pullRequest.headBranch,
-      });
-    }).pipe(
-      Effect.catch((error) =>
-        Effect.logWarning(
-          `GitManager.configurePullRequestHeadUpstream: failed to configure upstream for ${localBranch} -> ${pullRequest.headBranch} in ${cwd}: ${error.message}`,
-        ).pipe(Effect.asVoid),
-      ),
-    );
-
-  const materializePullRequestHeadBranch = (
-    cwd: string,
-    pullRequest: ResolvedPullRequest & PullRequestHeadRemoteInfo,
-    localBranch = pullRequest.headBranch,
-  ) =>
-    Effect.gen(function* () {
-      const repositoryNameWithOwner = resolveHeadRepositoryNameWithOwner(pullRequest) ?? "";
-
-      if (repositoryNameWithOwner.length === 0) {
-        yield* gitCore.fetchPullRequestBranch({
-          cwd,
-          prNumber: pullRequest.number,
-          branch: localBranch,
-        });
-        return;
-      }
-
-      const cloneUrls = yield* gitHubCli.getRepositoryCloneUrls({
-        cwd,
-        repository: repositoryNameWithOwner,
-      });
-      const originRemoteUrl = yield* gitCore.readConfigValue(cwd, "remote.origin.url");
-      const remoteUrl = shouldPreferSshRemote(originRemoteUrl) ? cloneUrls.sshUrl : cloneUrls.url;
-      const preferredRemoteName =
-        pullRequest.headRepositoryOwnerLogin?.trim() ||
-        repositoryNameWithOwner.split("/")[0]?.trim() ||
-        "fork";
-      const remoteName = yield* gitCore.ensureRemote({
-        cwd,
-        preferredName: preferredRemoteName,
-        url: remoteUrl,
-      });
-
-      yield* gitCore.fetchRemoteBranch({
-        cwd,
-        remoteName,
-        remoteBranch: pullRequest.headBranch,
-        localBranch,
-      });
-      yield* gitCore.setBranchUpstream({
-        cwd,
-        branch: localBranch,
-        remoteName,
-        remoteBranch: pullRequest.headBranch,
-      });
-    }).pipe(
-      Effect.catch(() =>
-        gitCore.fetchPullRequestBranch({
-          cwd,
-          prNumber: pullRequest.number,
-          branch: localBranch,
-        }),
-      ),
-    );
-  const fileSystem = yield* FileSystem.FileSystem;
-  const path = yield* Path.Path;
-
-  const tempDir = process.env.TMPDIR ?? process.env.TEMP ?? process.env.TMP ?? "/tmp";
-
-  const readConfigValueNullable = (cwd: string, key: string) =>
-    gitCore.readConfigValue(cwd, key).pipe(Effect.catch(() => Effect.succeed(null)));
-
-  const resolveCurrentUpstream = (cwd: string) =>
-    gitCore
-      .execute({
-        operation: "GitManager.resolveCurrentUpstream",
-        cwd,
-        args: ["rev-parse", "--abbrev-ref", "--symbolic-full-name", "@{upstream}"],
-        allowNonZeroExit: true,
-      })
-      .pipe(
-        Effect.map((result) => {
-          if (result.code !== 0) {
-            return null;
-          }
-
-          const upstreamRef = result.stdout.trim();
-          if (upstreamRef.length === 0 || upstreamRef === "@{upstream}") {
-            return null;
-          }
-
-          const separatorIndex = upstreamRef.indexOf("/");
-          if (separatorIndex <= 0) {
-            return null;
-          }
-
-          const remoteName = upstreamRef.slice(0, separatorIndex);
-          const upstreamBranch = upstreamRef.slice(separatorIndex + 1);
-          if (remoteName.length === 0 || upstreamBranch.length === 0) {
-            return null;
-          }
-
-          return {
-            upstreamRef,
-            remoteName,
-            upstreamBranch,
-          };
-        }),
-      );
-
-  const resolvePrimaryRemoteName = (cwd: string) =>
-    gitCore
-      .execute({
-        operation: "GitManager.resolvePrimaryRemoteName",
-        cwd,
-        args: ["remote"],
-        allowNonZeroExit: true,
-      })
-      .pipe(
-        Effect.map((result) =>
-          result.code === 0 ? selectPrimaryRemoteName(parseRemoteNames(result.stdout)) : null,
-        ),
-      );
-
-  const resolvePublishRemoteNameForBranch = (cwd: string, branch: string) =>
-    Effect.gen(function* () {
-      const branchPushRemote = yield* readConfigValueNullable(cwd, `branch.${branch}.pushRemote`);
-      const pushDefaultRemote = yield* readConfigValueNullable(cwd, "remote.pushDefault");
-      const primaryRemoteName = yield* resolvePrimaryRemoteName(cwd).pipe(
-        Effect.catch(() => Effect.succeed(null)),
-      );
-
-      return resolvePublishRemoteName({
-        branchPushRemote,
-        pushDefaultRemote,
-        primaryRemoteName,
-      });
-    });
-
-  const ensurePullAllowed = (cwd: string) =>
-    Effect.gen(function* () {
-      const currentUpstream = yield* resolveCurrentUpstream(cwd).pipe(
-        Effect.catch(() => Effect.succeed(null)),
-      );
-      if (currentUpstream && isProtectedRemoteName(currentUpstream.remoteName)) {
-        return yield* gitManagerError("pull", buildBlockedPullDetail());
-      }
-    });
-
-  const ensurePushAllowed = (cwd: string, fallbackBranch: string | null) =>
-    Effect.gen(function* () {
-      const details = yield* gitCore.statusDetails(cwd);
-      const branch = details.branch ?? fallbackBranch;
-      if (!branch) {
-        return;
-      }
-
-      if (details.hasUpstream) {
-        const currentUpstream = yield* resolveCurrentUpstream(cwd).pipe(
-          Effect.catch(() => Effect.succeed(null)),
-        );
-        if (currentUpstream && isProtectedRemoteName(currentUpstream.remoteName)) {
-          return yield* gitManagerError("push", buildBlockedPushDetail());
-        }
-        return;
-      }
-
-      const publishRemoteName = yield* resolvePublishRemoteNameForBranch(cwd, branch);
-      if (publishRemoteName && isProtectedRemoteName(publishRemoteName)) {
-        return yield* gitManagerError("push", buildBlockedPushDetail());
-      }
-    });
-
-  const pushCurrentBranch = (cwd: string, fallbackBranch: string | null) =>
-    Effect.gen(function* () {
-      yield* ensurePushAllowed(cwd, fallbackBranch);
-      return yield* gitCore.pushCurrentBranch(cwd, fallbackBranch);
-    });
-
-  const pullCurrentBranch = (cwd: string) =>
-    Effect.gen(function* () {
-      yield* ensurePullAllowed(cwd);
-      return yield* gitCore.pullCurrentBranch(cwd);
-    });
-
-  const resolveRemoteRepositoryContext = (cwd: string, remoteName: string | null) =>
-    Effect.gen(function* () {
-      if (!remoteName || isProtectedRemoteName(remoteName)) {
-        return resolveRepositoryContext(null);
-      }
-
-      const remoteUrl = yield* readConfigValueNullable(cwd, `remote.${remoteName}.url`);
-      return resolveRepositoryContext(
-        remoteUrl ? parseGitHubRepoSelector(remoteUrl, DEFAULT_GITHUB_HOSTNAME) : null,
-      );
-    });
-
-  const resolveBranchHeadContext = (
-    cwd: string,
-    details: { branch: string; upstreamRef: string | null },
-    baseRepositorySelector: string | null,
-  ) =>
-    Effect.gen(function* () {
-      const configuredRemoteName = yield* readConfigValueNullable(
-        cwd,
-        `branch.${details.branch}.remote`,
-      );
-      const remoteName = isProtectedRemoteName(configuredRemoteName) ? null : configuredRemoteName;
-      const headBranchFromUpstream = details.upstreamRef
-        ? extractBranchFromRef(details.upstreamRef)
-        : "";
-      const headBranch =
-        headBranchFromUpstream.length > 0 ? headBranchFromUpstream : details.branch;
-
-      const [remoteRepository, originRepository] = yield* Effect.all(
-        [
-          resolveRemoteRepositoryContext(cwd, remoteName),
-          resolveRemoteRepositoryContext(cwd, "origin"),
-        ],
-        { concurrency: "unbounded" },
-      );
-      const baseRepository =
-        baseRepositorySelector !== null
-          ? resolveRepositoryContext(baseRepositorySelector)
-          : originRepository;
-
-      const isCrossRepository =
-        remoteRepository.repositoryNameWithOwner !== null &&
-        baseRepository.repositoryNameWithOwner !== null
-          ? remoteRepository.repositoryNameWithOwner.toLowerCase() !==
-            baseRepository.repositoryNameWithOwner.toLowerCase()
-          : remoteName !== null &&
-            remoteName !== "origin" &&
-            remoteRepository.repositoryNameWithOwner !== null;
-
-      const ownerHeadSelector =
-        remoteRepository.ownerLogin && headBranch.length > 0
-          ? `${remoteRepository.ownerLogin}:${headBranch}`
-          : null;
-      const remoteAliasHeadSelector =
-        remoteName && headBranch.length > 0 ? `${remoteName}:${headBranch}` : null;
-      const shouldProbeRemoteOwnedSelectors =
-        isCrossRepository || (remoteName !== null && remoteName !== "origin");
-
-      const headSelectors: string[] = [];
-      if (isCrossRepository && shouldProbeRemoteOwnedSelectors) {
-        appendUnique(headSelectors, ownerHeadSelector);
-        appendUnique(
-          headSelectors,
-          remoteAliasHeadSelector !== ownerHeadSelector ? remoteAliasHeadSelector : null,
-        );
-      }
-      appendUnique(headSelectors, details.branch);
-      appendUnique(headSelectors, headBranch !== details.branch ? headBranch : null);
-      if (!isCrossRepository && shouldProbeRemoteOwnedSelectors) {
-        appendUnique(headSelectors, ownerHeadSelector);
-        appendUnique(
-          headSelectors,
-          remoteAliasHeadSelector !== ownerHeadSelector ? remoteAliasHeadSelector : null,
-        );
-      }
-
-      return {
-        localBranch: details.branch,
-        headBranch,
-        headSelectors,
-        preferredHeadSelector:
-          ownerHeadSelector && isCrossRepository ? ownerHeadSelector : headBranch,
-        remoteName,
-        headRepositoryNameWithOwner: remoteRepository.repositoryNameWithOwner,
-        headRepositoryOwnerLogin: remoteRepository.ownerLogin,
-        isCrossRepository,
-      } satisfies BranchHeadContext;
-    });
-
-  const findOpenPr = (cwd: string, headSelectors: ReadonlyArray<string>) =>
-    Effect.gen(function* () {
-      const repositorySelector = yield* resolveRepositorySelector(cwd);
-
-      for (const headSelector of headSelectors) {
-        const pullRequests = yield* gitHubCli.listOpenPullRequests(
-          withOptionalRepo({ cwd, headSelector, limit: 1 }, repositorySelector),
-        );
-
-        const [firstPullRequest] = pullRequests;
-        if (firstPullRequest) {
-          return {
-            number: firstPullRequest.number,
-            title: firstPullRequest.title,
-            url: firstPullRequest.url,
-            baseRefName: firstPullRequest.baseRefName,
-            headRefName: firstPullRequest.headRefName,
-            state: "open",
-            updatedAt: null,
-          } satisfies PullRequestInfo;
-        }
-      }
-
-      return null;
-    });
-
-  const findLatestPr = (cwd: string, details: { branch: string; upstreamRef: string | null }) =>
-    Effect.gen(function* () {
-      const repositorySelector = yield* resolveRepositorySelector(cwd);
-      const headContext = yield* resolveBranchHeadContext(cwd, details, repositorySelector);
-      const parsedByNumber = new Map<number, PullRequestInfo>();
-
-      for (const headSelector of headContext.headSelectors) {
-        const stdout = yield* gitHubCli
-          .execute({
-            cwd,
-            args: [
-              "pr",
-              "list",
-              ...(repositorySelector ? ["--repo", repositorySelector] : []),
-              "--head",
-              headSelector,
-              "--state",
-              "all",
-              "--limit",
-              "20",
-              "--json",
-              "number,title,url,baseRefName,headRefName,state,mergedAt,updatedAt",
-            ],
-          })
-          .pipe(Effect.map((result) => result.stdout));
-
-        const raw = stdout.trim();
-        if (raw.length === 0) {
-          continue;
-        }
-
-        const parsedJson = yield* Effect.try({
-          try: () => JSON.parse(raw) as unknown,
-          catch: (cause) =>
-            gitManagerError("findLatestPr", "GitHub CLI returned invalid PR list JSON.", cause),
-        });
-
-        for (const pr of parsePullRequestList(parsedJson)) {
-          parsedByNumber.set(pr.number, pr);
-        }
-      }
-
-      const parsed = Array.from(parsedByNumber.values()).toSorted((a, b) => {
-        const left = a.updatedAt ? Date.parse(a.updatedAt) : 0;
-        const right = b.updatedAt ? Date.parse(b.updatedAt) : 0;
-        return right - left;
-      });
-
-      const latestOpenPr = parsed.find((pr) => pr.state === "open");
-      if (latestOpenPr) {
-        return latestOpenPr;
-      }
-      return parsed[0] ?? null;
-    });
-
-  const resolveBaseBranch = (
-    cwd: string,
-    branch: string,
-    upstreamRef: string | null,
-    headContext: Pick<BranchHeadContext, "isCrossRepository">,
-    repositorySelector: string | null,
-  ) =>
-    Effect.gen(function* () {
-      const configured = yield* gitCore.readConfigValue(cwd, `branch.${branch}.gh-merge-base`);
-      if (configured) return configured;
-
-      if (upstreamRef && !headContext.isCrossRepository) {
-        const upstreamBranch = extractBranchFromRef(upstreamRef);
-        if (upstreamBranch.length > 0 && upstreamBranch !== branch) {
-          return upstreamBranch;
-        }
-      }
-
-      const defaultFromGh = yield* gitHubCli
-        .getDefaultBranch(withOptionalRepo({ cwd }, repositorySelector))
-        .pipe(Effect.catch(() => Effect.succeed(null)));
-      if (defaultFromGh) {
-        return defaultFromGh;
-      }
-
-      return "main";
-    });
-
-  const resolveCommitAndBranchSuggestion = (input: {
-    cwd: string;
-    branch: string | null;
-    commitMessage?: string;
-    /** When true, also produce a semantic feature branch name. */
-    includeBranch?: boolean;
-    filePaths?: readonly string[];
-    modelSelection: ModelSelection;
-  }) =>
-    Effect.gen(function* () {
-      const context = yield* gitCore.prepareCommitContext(input.cwd, input.filePaths);
-      if (!context) {
-        return null;
-      }
-
-      const customCommit = parseCustomCommitMessage(input.commitMessage ?? "");
-      if (customCommit) {
-        return {
-          subject: customCommit.subject,
-          body: customCommit.body,
-          ...(input.includeBranch
-            ? { branch: sanitizeFeatureBranchName(customCommit.subject) }
-            : {}),
-          commitMessage: formatCommitMessage(customCommit.subject, customCommit.body),
-        };
-      }
-
-      const generated = yield* textGeneration
-        .generateCommitMessage({
-          cwd: input.cwd,
-          branch: input.branch,
-          stagedSummary: limitContext(context.stagedSummary, 8_000),
-          stagedPatch: limitContext(context.stagedPatch, 50_000),
-          ...(input.includeBranch ? { includeBranch: true } : {}),
-          modelSelection: input.modelSelection,
-        })
-        .pipe(Effect.map((result) => sanitizeCommitMessage(result)));
-
-      return {
-        subject: generated.subject,
-        body: generated.body,
-        ...(generated.branch !== undefined ? { branch: generated.branch } : {}),
-        commitMessage: formatCommitMessage(generated.subject, generated.body),
-      };
-    });
-
-  const runCommitStep = (
-    modelSelection: ModelSelection,
-    cwd: string,
-    branch: string | null,
-    commitMessage?: string,
-    preResolvedSuggestion?: CommitAndBranchSuggestion,
-    filePaths?: readonly string[],
-    _progressReporter?: GitActionProgressReporter,
-    _actionId?: string,
-  ) =>
-    Effect.gen(function* () {
-      const suggestion =
-        preResolvedSuggestion ??
-        (yield* resolveCommitAndBranchSuggestion({
-          cwd,
-          branch,
-          ...(commitMessage ? { commitMessage } : {}),
-          ...(filePaths ? { filePaths } : {}),
-          modelSelection,
-        }));
-      if (!suggestion) {
-        return { status: "skipped_no_changes" as const };
-      }
-
-      const { commitSha } = yield* gitCore.commit(cwd, suggestion.subject, suggestion.body);
-      return {
-        status: "created" as const,
-        commitSha,
-        subject: suggestion.subject,
-      };
-    });
-
-  const runPrStep = (
-    modelSelection: ModelSelection,
-    cwd: string,
-    fallbackBranch: string | null,
-    issueLink: GitHubIssueLink | null,
-  ) =>
-    Effect.gen(function* () {
-      const details = yield* gitCore.statusDetails(cwd);
-      const branch = details.branch ?? fallbackBranch;
-      if (!branch) {
-        return yield* gitManagerError(
-          "runPrStep",
-          "Cannot create a pull request from detached HEAD.",
-        );
-      }
-      if (!details.hasUpstream) {
-        return yield* gitManagerError(
-          "runPrStep",
-          "Current branch has not been pushed. Push before creating a PR.",
-        );
-      }
-
-      const repositorySelector = yield* resolveRepositorySelector(cwd);
-      const headContext = yield* resolveBranchHeadContext(
-        cwd,
-        {
-          branch,
-          upstreamRef: details.upstreamRef,
-        },
-        repositorySelector,
-      );
-
-      const existing = yield* findOpenPr(cwd, headContext.headSelectors);
-      if (existing) {
-        return {
-          status: "opened_existing" as const,
-          url: existing.url,
-          number: existing.number,
-          baseBranch: existing.baseRefName,
-          headBranch: existing.headRefName,
-          title: existing.title,
-        };
-      }
-
-      const baseBranch = yield* resolveBaseBranch(
-        cwd,
-        branch,
-        details.upstreamRef,
-        headContext,
-        repositorySelector,
-      );
-      const rangeContext = yield* gitCore.readRangeContext(cwd, baseBranch);
-
-      const generated = yield* textGeneration.generatePrContent({
-        cwd,
-        baseBranch,
-        headBranch: headContext.headBranch,
-        commitSummary: limitContext(rangeContext.commitSummary, 20_000),
-        diffSummary: limitContext(rangeContext.diffSummary, 20_000),
-        diffPatch: limitContext(rangeContext.diffPatch, 60_000),
-        modelSelection,
-      });
-      const body = appendIssueClosingFooter(generated.body, issueLink);
-
-      const bodyFile = path.join(tempDir, `t3code-pr-body-${process.pid}-${randomUUID()}.md`);
-      yield* fileSystem
-        .writeFileString(bodyFile, body)
-        .pipe(
-          Effect.mapError((cause) =>
-            gitManagerError("runPrStep", "Failed to write pull request body temp file.", cause),
-          ),
-        );
-      yield* gitHubCli
-        .createPullRequest(
-          withOptionalRepo(
-            {
-              cwd,
-              baseBranch,
-              headSelector: headContext.preferredHeadSelector,
-              title: generated.title,
-              bodyFile,
-            },
-            repositorySelector,
-          ),
-        )
-        .pipe(Effect.ensuring(fileSystem.remove(bodyFile).pipe(Effect.catch(() => Effect.void))));
-
-      const created = yield* findOpenPr(cwd, headContext.headSelectors);
-      if (!created) {
-        return {
-          status: "created" as const,
-          baseBranch,
-          headBranch: headContext.headBranch,
-          title: generated.title,
-        };
-      }
-
-      return {
-        status: "created" as const,
-        url: created.url,
-        number: created.number,
-        baseBranch: created.baseRefName,
-        headBranch: created.headRefName,
-        title: created.title,
-      };
-    });
-
-  const status: GitManagerShape["status"] = Effect.fnUntraced(function* (input) {
-    const details = yield* gitCore.statusDetails(input.cwd);
-
-    const pr =
-      details.branch !== null
-        ? yield* findLatestPr(input.cwd, {
-            branch: details.branch,
-            upstreamRef: details.upstreamRef,
-          }).pipe(
-            Effect.map((latest) => (latest ? toStatusPr(latest) : null)),
-            Effect.catch(() => Effect.succeed(null)),
-          )
-        : null;
-
-    return {
-      branch: details.branch,
-      hasWorkingTreeChanges: details.hasWorkingTreeChanges,
-      workingTree: details.workingTree,
-      hasUpstream: details.hasUpstream,
-      aheadCount: details.aheadCount,
-      behindCount: details.behindCount,
-      merge: details.merge,
-      pr,
-    };
-  });
-
-  const pull: GitManagerShape["pull"] = Effect.fnUntraced(function* (input) {
-    return yield* pullCurrentBranch(input.cwd);
-  });
-
-  const repositoryContext: GitManagerShape["repositoryContext"] = Effect.fnUntraced(
-    function* (input) {
-      return yield* gitCore.getRepositoryContext(input.cwd);
-    },
-  );
-
-  const resolvePullRequest: GitManagerShape["resolvePullRequest"] = Effect.fnUntraced(
-    function* (input) {
-      const repositorySelector = yield* resolveRepositorySelector(input.cwd);
-      const pullRequest = yield* gitHubCli
-        .getPullRequest(
-          withOptionalRepo(
-            {
-              cwd: input.cwd,
-              reference: normalizePullRequestReference(input.reference),
-            },
-            repositorySelector,
-          ),
-        )
-        .pipe(Effect.map((resolved) => toResolvedPullRequest(resolved)));
-
-      return { pullRequest };
-    },
-  );
-
-  const preparePullRequestThread: GitManagerShape["preparePullRequestThread"] = Effect.fnUntraced(
-    function* (input) {
-      const normalizedReference = normalizePullRequestReference(input.reference);
-      const rootWorktreePath = canonicalizeExistingPath(input.cwd);
-      const repositorySelector = yield* resolveRepositorySelector(input.cwd);
-      const pullRequestSummary = yield* gitHubCli.getPullRequest(
-        withOptionalRepo(
-          {
-            cwd: input.cwd,
-            reference: normalizedReference,
-          },
-          repositorySelector,
-        ),
-      );
-      const pullRequest = toResolvedPullRequest(pullRequestSummary);
-
-      if (input.mode === "local") {
-        yield* gitHubCli.checkoutPullRequest(
-          withOptionalRepo(
-            {
-              cwd: input.cwd,
-              reference: normalizedReference,
-              force: true,
-            },
-            repositorySelector,
-          ),
-        );
-        const details = yield* gitCore.statusDetails(input.cwd);
-        yield* configurePullRequestHeadUpstream(
-          input.cwd,
-          {
-            ...pullRequest,
-            ...toPullRequestHeadRemoteInfo(pullRequestSummary),
-          },
-          details.branch ?? pullRequest.headBranch,
-        );
-        return {
-          pullRequest,
-          branch: details.branch ?? pullRequest.headBranch,
-          worktreePath: null,
-        };
-      }
-
-      const ensureExistingWorktreeUpstream = (worktreePath: string) =>
-        Effect.gen(function* () {
-          const details = yield* gitCore.statusDetails(worktreePath);
-          yield* configurePullRequestHeadUpstream(
-            worktreePath,
-            {
-              ...pullRequest,
-              ...toPullRequestHeadRemoteInfo(pullRequestSummary),
-            },
-            details.branch ?? pullRequest.headBranch,
-          );
-        });
-
-      const pullRequestWithRemoteInfo = {
-        ...pullRequest,
-        ...toPullRequestHeadRemoteInfo(pullRequestSummary),
-      } as const;
-      const localPullRequestBranch =
-        resolvePullRequestWorktreeLocalBranchName(pullRequestWithRemoteInfo);
-
-      const findLocalHeadBranch = (cwd: string) =>
-        gitCore.listBranches({ cwd }).pipe(
-          Effect.map((result) => {
-            const localBranch = result.branches.find(
-              (branch) => !branch.isRemote && branch.name === localPullRequestBranch,
-            );
-            if (localBranch) {
-              return localBranch;
-            }
-            if (localPullRequestBranch === pullRequest.headBranch) {
-              return null;
-            }
-            return (
-              result.branches.find(
-                (branch) =>
-                  !branch.isRemote &&
-                  branch.name === pullRequest.headBranch &&
-                  branch.worktreePath !== null &&
-                  canonicalizeExistingPath(branch.worktreePath) !== rootWorktreePath,
-              ) ?? null
-            );
-          }),
-        );
-
-      const existingBranchBeforeFetch = yield* findLocalHeadBranch(input.cwd);
-      const existingBranchBeforeFetchPath = existingBranchBeforeFetch?.worktreePath
-        ? canonicalizeExistingPath(existingBranchBeforeFetch.worktreePath)
-        : null;
-      if (
-        existingBranchBeforeFetch?.worktreePath &&
-        existingBranchBeforeFetchPath !== rootWorktreePath
-      ) {
-        yield* ensureExistingWorktreeUpstream(existingBranchBeforeFetch.worktreePath);
-        return {
-          pullRequest,
-          branch: localPullRequestBranch,
-          worktreePath: existingBranchBeforeFetch.worktreePath,
-        };
-      }
-      if (existingBranchBeforeFetchPath === rootWorktreePath) {
-        return yield* gitManagerError(
-          "preparePullRequestThread",
-          "This PR branch is already checked out in the main repo. Use Local, or switch the main repo off that branch before creating a worktree thread.",
-        );
-      }
-
-      yield* materializePullRequestHeadBranch(
-        input.cwd,
-        pullRequestWithRemoteInfo,
-        localPullRequestBranch,
-      );
-
-      const existingBranchAfterFetch = yield* findLocalHeadBranch(input.cwd);
-      const existingBranchAfterFetchPath = existingBranchAfterFetch?.worktreePath
-        ? canonicalizeExistingPath(existingBranchAfterFetch.worktreePath)
-        : null;
-      if (
-        existingBranchAfterFetch?.worktreePath &&
-        existingBranchAfterFetchPath !== rootWorktreePath
-      ) {
-        yield* ensureExistingWorktreeUpstream(existingBranchAfterFetch.worktreePath);
-        return {
-          pullRequest,
-          branch: localPullRequestBranch,
-          worktreePath: existingBranchAfterFetch.worktreePath,
-        };
-      }
-      if (existingBranchAfterFetchPath === rootWorktreePath) {
-        return yield* gitManagerError(
-          "preparePullRequestThread",
-          "This PR branch is already checked out in the main repo. Use Local, or switch the main repo off that branch before creating a worktree thread.",
-        );
-      }
-
-      const worktree = yield* gitCore.createWorktree({
-        cwd: input.cwd,
-        branch: localPullRequestBranch,
-        path: null,
-      });
-      yield* ensureExistingWorktreeUpstream(worktree.worktree.path);
-
-      return {
-        pullRequest,
-        branch: worktree.worktree.branch,
-        worktreePath: worktree.worktree.path,
-      };
-    },
-  );
-
-  const runFeatureBranchStep = (
-    modelSelection: ModelSelection,
-    cwd: string,
-    branch: string | null,
-    commitMessage?: string,
-    filePaths?: readonly string[],
-  ) =>
-    Effect.gen(function* () {
-      const suggestion = yield* resolveCommitAndBranchSuggestion({
-        cwd,
-        branch,
-        ...(commitMessage ? { commitMessage } : {}),
-        ...(filePaths ? { filePaths } : {}),
-        includeBranch: true,
-        modelSelection,
-      });
-      if (!suggestion) {
-        return yield* gitManagerError(
-          "runFeatureBranchStep",
-          "Cannot create a feature branch because there are no changes to commit.",
-        );
-      }
-
-      const preferredBranch = suggestion.branch ?? sanitizeFeatureBranchName(suggestion.subject);
-      const existingBranchNames = yield* gitCore.listLocalBranchNames(cwd);
-      const resolvedBranch = resolveAutoFeatureBranchName(existingBranchNames, preferredBranch);
-
-      yield* gitCore.createBranch({ cwd, branch: resolvedBranch });
-      yield* Effect.scoped(gitCore.checkoutBranch({ cwd, branch: resolvedBranch }));
-
-      return {
-        branchStep: { status: "created" as const, name: resolvedBranch },
-        resolvedCommitMessage: suggestion.commitMessage,
-        resolvedCommitSuggestion: suggestion,
-      };
-    });
-
-  const runPromoteStep = (
-    cwd: string,
-    sourceBranch: string,
-    targetBranch: string,
-    modelSelection: ModelSelection,
-    commitMessage?: string,
-    filePaths?: readonly string[],
-  ) =>
-    Effect.gen(function* () {
-      const commit = yield* runCommitStep(
-        modelSelection,
-        cwd,
-        sourceBranch,
-        commitMessage,
-        undefined,
-        filePaths,
-      );
-      const sourceCommitAndPush = {
-        commit,
-        push: yield* gitCore.pushCurrentBranch(cwd, sourceBranch),
-      };
-
-      // Step 3: Checkout target branch
-      yield* Effect.scoped(gitCore.checkoutBranch({ cwd, branch: targetBranch }));
-
-      // Step 4: Pull latest target (to avoid push conflicts)
-      yield* pullCurrentBranch(cwd).pipe(Effect.catch(() => Effect.void));
-
-      // Step 5: Merge feature branch into target
-      const mergeResult = yield* gitCore.mergeBranches({
-        cwd,
-        sourceBranch,
-        targetBranch,
-      });
-
-      if (mergeResult.status === "conflicted") {
-        // Leave in conflict state for user to resolve
-        return {
-          commit: sourceCommitAndPush.commit,
-          push: sourceCommitAndPush.push,
-          promote: {
-            status: "conflicts" as const,
-            sourceBranch,
-            targetBranch,
-            conflictedFiles: mergeResult.conflictedFiles,
-            branchDeleted: false,
-          },
-        };
-      }
-
-      // Step 6: Push target to origin
-      const targetPush = yield* pushCurrentBranch(cwd, targetBranch);
-
-      // Step 7: Delete feature branch (local + remote)
-      const deleteResult = yield* gitCore.deleteBranch({
-        cwd,
-        branch: sourceBranch,
-        deleteRemote: true,
-      });
-
-      return {
-        commit: sourceCommitAndPush.commit,
-        push: targetPush,
-        promote: {
-          status: "promoted" as const,
-          sourceBranch,
-          targetBranch,
-          branchDeleted: deleteResult.deletedLocal || deleteResult.deletedRemote,
-        },
-      };
-    });
-
-  const runStackedAction: GitManagerShape["runStackedAction"] = Effect.fnUntraced(
-    function* (input, options) {
-      const progress = createProgressEmitter(input, options);
-      const wantsPush = input.action !== "commit";
-      const wantsPr = input.action === "commit_push_pr";
-      const wantsPromote = input.action === "promote";
-      const phases: GitActionProgressPhase[] = [
-        ...(input.featureBranch ? (["branch"] as const) : []),
-        "commit",
-        ...(wantsPush ? (["push"] as const) : []),
-        ...(wantsPr ? (["pr"] as const) : []),
-      ];
-      let currentPhase: GitActionProgressPhase | null = null;
-
-      const runAction = Effect.gen(function* () {
-        yield* progress.emit({
-          kind: "action_started",
-          phases,
-        });
-
-        const initialStatus = yield* gitCore.statusDetails(input.cwd);
-        if (!input.featureBranch && wantsPush && !initialStatus.branch) {
-          return yield* gitManagerError("runStackedAction", "Cannot push from detached HEAD.");
-        }
-        if (!input.featureBranch && wantsPr && !initialStatus.branch) {
-          return yield* gitManagerError(
-            "runStackedAction",
-            "Cannot create a pull request from detached HEAD.",
-          );
-        }
-        if (wantsPromote && !initialStatus.branch) {
-          return yield* gitManagerError("runStackedAction", "Cannot promote from detached HEAD.");
-        }
-        if (wantsPromote && !input.targetBranch) {
-          return yield* gitManagerError(
-            "runStackedAction",
-            "Promote action requires a target branch.",
-          );
-        }
-
-        const modelSelection = yield* serverSettingsService.getSettings.pipe(
-          Effect.map((settings) => settings.textGenerationModelSelection),
-          Effect.mapError((cause) =>
-            gitManagerError("runStackedAction", "Failed to get server settings.", cause),
-          ),
-        );
-
-        if (wantsPromote && initialStatus.branch && input.targetBranch) {
-          currentPhase = "commit";
-          const promoteResult = yield* runPromoteStep(
-            input.cwd,
-            initialStatus.branch,
-            input.targetBranch,
-            modelSelection,
-            input.commitMessage,
-            input.filePaths,
-          );
-
-          const result = {
-            action: input.action,
-            branch: { status: "skipped_not_requested" as const },
-            commit: promoteResult.commit,
-            push: promoteResult.push,
-            pr: { status: "skipped_not_requested" as const },
-            promote: promoteResult.promote,
-          };
-          yield* progress.emit({
-            kind: "action_finished",
-            result,
-          });
-          return result;
-        }
-
-        let branchStep: { status: "created" | "skipped_not_requested"; name?: string };
-        let commitMessageForStep = input.commitMessage;
-        let preResolvedCommitSuggestion: CommitAndBranchSuggestion | undefined = undefined;
-
-        if (input.featureBranch) {
-          currentPhase = "branch";
-          yield* progress.emit({
-            kind: "phase_started",
-            phase: "branch",
-            label: "Preparing feature branch...",
-          });
-          const result = yield* runFeatureBranchStep(
-            modelSelection,
-            input.cwd,
-            initialStatus.branch,
-            input.commitMessage,
-            input.filePaths,
-          );
-          branchStep = result.branchStep;
-          commitMessageForStep = result.resolvedCommitMessage;
-          preResolvedCommitSuggestion = result.resolvedCommitSuggestion;
-        } else {
-          branchStep = { status: "skipped_not_requested" as const };
-        }
-
-        const currentBranch = branchStep.name ?? initialStatus.branch;
-
-        currentPhase = "commit";
-        const commit = yield* runCommitStep(
-          modelSelection,
-          input.cwd,
-          currentBranch,
-          commitMessageForStep,
-          preResolvedCommitSuggestion,
-          input.filePaths,
-          options?.progressReporter,
-          progress.actionId,
-        );
-
-        const push = wantsPush
-          ? yield* progress
-              .emit({
-                kind: "phase_started",
-                phase: "push",
-                label: "Pushing...",
-              })
-              .pipe(
-                Effect.flatMap(() =>
-                  Effect.gen(function* () {
-                    currentPhase = "push";
-                    return yield* gitCore.pushCurrentBranch(input.cwd, currentBranch);
-                  }),
-                ),
-              )
-          : { status: "skipped_not_requested" as const };
-
-        const pr = wantsPr
-          ? yield* progress
-              .emit({
-                kind: "phase_started",
-                phase: "pr",
-                label: "Creating PR...",
-              })
-              .pipe(
-                Effect.flatMap(() =>
-                  Effect.gen(function* () {
-                    currentPhase = "pr";
-                    return yield* runPrStep(
-                      modelSelection,
-                      input.cwd,
-                      currentBranch,
-                      input.issueLink ?? null,
-                    );
-                  }),
-                ),
-              )
-          : { status: "skipped_not_requested" as const };
-
-        const result = {
-          action: input.action,
-          branch: branchStep,
-          commit,
-          push,
-          pr,
-          promote: { status: "skipped_not_requested" as const },
-        };
-        yield* progress.emit({
-          kind: "action_finished",
-          result,
-        });
-        return result;
-      });
-
-      return yield* runAction.pipe(
-        Effect.catch((error) =>
-          progress
-            .emit({
-              kind: "action_failed",
-              phase: currentPhase,
-              message: error instanceof Error ? error.message : String(error),
-            })
-            .pipe(Effect.flatMap(() => Effect.fail(error))),
-        ),
-      );
-    },
-  );
+  const forkGitDomain = yield* ForkGitDomain;
 
   return {
-    status,
-    pull,
-    repositoryContext,
-    resolvePullRequest,
-    preparePullRequestThread,
-    runStackedAction,
+    status: (input) => forkGitDomain.status(input),
+    pull: (input) => forkGitDomain.pull(input),
+    repositoryContext: (input) => forkGitDomain.repositoryContext(input),
+    resolvePullRequest: (input) => forkGitDomain.resolvePullRequest(input),
+    preparePullRequestThread: (input) => forkGitDomain.preparePullRequestThread(input),
+    runStackedAction: (input, options) => forkGitDomain.runStackedAction(input, options),
   } satisfies GitManagerShape;
 });
