+++ apps/server/src/serverLayers.ts
@@ -28,6 +28,7 @@ import { ServerSettingsService } from "./serverSettings";
 
 import { TerminalManagerLive } from "./terminal/Layers/Manager";
 import { KeybindingsLive } from "./keybindings";
+import { ForkGitDomainLive } from "./git/Layers/ForkGitDomain";
 import { GitManagerLive } from "./git/Layers/GitManager";
 import { GitCoreLive } from "./git/Layers/GitCore";
 import { GitServiceLive } from "./git/Layers/GitService";
@@ -133,11 +134,12 @@ export function makeServerRuntimeServicesLayer() {
 
   const terminalLayer = TerminalManagerLive.pipe(Layer.provide(makeRuntimePtyAdapterLayer()));
 
-  const gitManagerLayer = GitManagerLive.pipe(
+  const forkGitDomainLayer = ForkGitDomainLive.pipe(
     Layer.provideMerge(gitCoreLayer),
     Layer.provideMerge(GitHubCliLive),
     Layer.provideMerge(textGenerationLayer),
   );
+  const gitManagerLayer = GitManagerLive.pipe(Layer.provideMerge(forkGitDomainLayer));
   const gitHubManagerLayer = GitHubManagerLive.pipe(
     Layer.provideMerge(gitCoreLayer),
     Layer.provideMerge(GitHubCliLive),
